Refactored Code

+++ code/jvm/src/main/kotlin/daw/isel/pt/gomoku/services/LobbyServices.kt
@@ -2,19 +2,22 @@ package daw.isel.pt.gomoku.services
 
 import daw.isel.pt.gomoku.domain.Lobby
 import daw.isel.pt.gomoku.repository.interfaces.transactions.TransactionManager
-import daw.isel.pt.gomoku.services.exceptions.AlreadyInLobbyException
-import daw.isel.pt.gomoku.services.exceptions.InvalidCredentialsException
-import daw.isel.pt.gomoku.services.exceptions.LobbyErrorMessages
-import daw.isel.pt.gomoku.services.exceptions.NotFoundException
+import daw.isel.pt.gomoku.services.exceptions.*
 import org.springframework.stereotype.Component
 
 @Component
 class LobbyServices(private val transactionManager: TransactionManager) {
     fun createLobby(userId: Int, name: String?): Lobby {
         return transactionManager.run {
-            if(name.isNullOrEmpty()) throw InvalidCredentialsException("Invalid name for lobby")
+            if(it.usersRepository.getUser(userId) == null)
+                throw NotFoundException(UserErrorMessages.USER_NOT_FOUND)
+            if(name.isNullOrEmpty())
+                throw InvalidCredentialsException(LobbyErrorMessages.INVALID_NAME)
             if(it.lobbyRepository.isNotInLobby(userId))
-                it.lobbyRepository.createLobby(userId, name)
+                it.lobbyRepository.createLobby(
+                    userId = userId,
+                    name = name
+                )
             else throw AlreadyInLobbyException(LobbyErrorMessages.USER_ALREADY_IN_LOBBY)
         }
     }
@@ -27,8 +30,13 @@ class LobbyServices(private val transactionManager: TransactionManager) {
 
     fun joinLobby(userId: Int, lobbyId: Int): Boolean {
         return transactionManager.run {
+            if(it.usersRepository.getUser(userId) == null)
+                throw NotFoundException(UserErrorMessages.USER_NOT_FOUND)
             if(it.lobbyRepository.isNotInLobby(userId))
-                it.lobbyRepository.joinLobby(lobbyId, userId)
+                it.lobbyRepository.joinLobby(
+                    lobbyId = lobbyId,
+                    userId = userId
+                )
             else throw AlreadyInLobbyException(LobbyErrorMessages.USER_ALREADY_IN_LOBBY)
         }
     }
@@ -46,15 +54,23 @@ class LobbyServices(private val transactionManager: TransactionManager) {
      */
     fun deleteLobby(userId: Int, lobbyId: Int): Boolean{
         return transactionManager.run {
+            if(it.usersRepository.getUser(userId) == null)
+                throw NotFoundException(UserErrorMessages.USER_NOT_FOUND)
             val lobby = it.lobbyRepository.getLobby(lobbyId)
             if(lobby != null) {
                 when(lobby.p1) {
                     userId -> {
                         when(lobby.p2) {
                             null ->  it.lobbyRepository.deleteLobby(lobbyId)
-                            else ->  it.lobbyRepository.switchLobbyAdmin(lobbyId, userId)
+                            else ->  it.lobbyRepository.switchLobbyAdmin(
+                                lobbyId = lobbyId,
+                                userId = userId
+                            )
                         }
-                    } else -> it.lobbyRepository.quitLobby(lobbyId, userId)
+                    } else -> it.lobbyRepository.quitLobby(
+                                    lobbyId = lobbyId,
+                                    userId = userId
+                            )
                 }
             } else throw NotFoundException(LobbyErrorMessages.LOBBY_NOT_FOUND)
         }
